Rename controls module and add comments
