+++ src/ui/renderer.mjs
@@ -1,0 +1,2 @@
+// Re-export preview renderer as the main renderer entry point
+export { frame } from './preview-renderer.mjs';
