+++ src/ui/subviews/readme.md
@@ -1,4 +1,4 @@
-# Controls
+# Subviews
 
 Dynamic UI widgets for effect parameters.
