+++ src/ui/subviews/speedSlider.mjs
